+++ apps/studio/components/layouts/ProjectLayout/ProjectLayout.tsx
@@ -34,7 +34,6 @@ import RestoreFailedState from './RestoreFailedState'
 import RestoringState from './RestoringState'
 import { UpgradingState } from './UpgradingState'
 import { useSheet } from 'components/ui/Sheet'
-import MobileViewNav from './NavigationBar/MobileViewNav'
 
 // [Joshen] This is temporary while we unblock users from managing their project
 // if their project is not responding well for any reason. Eventually needs a bit of an overhaul
@@ -152,9 +151,6 @@ const ProjectLayout = forwardRef<HTMLDivElement, PropsWithChildren<ProjectLayout
             {!hideIconBar && <NavigationBar />}
             {/* Top Nav to access products from mobile */}
             {!hideIconBar && <MobileNavigationBar />}
-            {showProductMenu && productMenu && (
-              <MobileViewNav title={product} productMenu={productMenu} />
-            )}
 
             {/* Product menu bar */}
             <ResizablePanelGroup
@@ -188,7 +184,9 @@ const ProjectLayout = forwardRef<HTMLDivElement, PropsWithChildren<ProjectLayout
                 disabled={resizableSidebar ? false : true}
               />
               <ResizablePanel id="panel-right" className="h-full flex flex-col">
-                {!hideHeader && IS_PLATFORM && <LayoutHeader />}
+                {!hideHeader && IS_PLATFORM && (
+                  <LayoutHeader showProductMenu={showProductMenu} productMenu={productMenu} />
+                )}
                 <ResizablePanelGroup
                   className="h-full w-full overflow-x-hidden flex-1"
                   direction="horizontal"
